# Thread loops over blocks