--- a/src/.gitignore
+++ b/src/.gitignore
@@ -3,3 +3,4 @@ makedeps.txt
 *.mod
 *.o
 datestamp.f90
+TAGS