--- a/src/functions_on_grid_module.f90
+++ b/src/functions_on_grid_module.f90
@@ -67,7 +67,7 @@ module functions_on_grid
   !!
   type fn_on_grid
 
-     real(double), dimension(:), pointer :: griddata
+     real(double), dimension(:), allocatable :: griddata
      integer :: size
      integer :: type