--- a/src/gha.make
+++ b/src/gha.make
@@ -1,5 +1,5 @@
 # Linking flags
-LINKFLAGS= -L/usr/lib -L/usr/lib/x86_64-linux-gnu
+LINKFLAGS=-fopenmp -L/usr/lib -L/usr/lib/x86_64-linux-gnu
 ARFLAGS=
 # Set BLAS and LAPACK libraries
 BLAS= -llapack -lblas
@@ -18,5 +18,5 @@ DIAG_DUMMY =
 LIBS= $(XC_LIB) -lscalapack-openmpi $(BLAS) $(FFT_LIB)
 # Compilation flags
 # NB for gcc10 you need to add -fallow-argument-mismatch
-COMPFLAGS= -O3 $(XC_COMPFLAGS) -fallow-argument-mismatch
+COMPFLAGS= -O3 -fopenmp $(XC_COMPFLAGS) -fallow-argument-mismatch
 COMPFLAGS_F77= $(COMPFLAGS)