--- a/src/odd.obj
+++ b/src/odd.obj
@@ -8,6 +8,7 @@ ODD_OBJS = global_module.o \
            logicals_module.o \
            memory_module.o \
            mpi.o \
+           omp_module.o \
            timer_module.o \
            timer_stdclocks_module.o \
            auxiliary_types.o \