PR to close #178 

This PR adds threading over the biggest loops over blocks in `calc_matrix_elements_module` and `PAO_grid_transform_module`.

In `calc_matrix_elements_module` threading is done in subroutines `get_matrix_elements_new` and `act_on_vectors_new`. Both subroutines have a deep loop nest where the loop over blocks is the outermost loop. In `get_matrix_elements_new` the innermost loop accumulates data in `send_array`, therefore a reduction is done at the end of the parallel region. In `get_matrix_elements_new`, `gridfunctions%griddata` is updated with a `gemm` call in the innermost loop. The updates are done in separate indices by each loop iteration, therefore it can be operated on by all threads in parallel.

In `PAO_grid_transform_module`, the `single_PAO_to_grid` subroutine is rewritten so that threading over blocks can be done. The innermost loop updates `gridfunctions%griddata`. However the index was calculated sequentially inside the loop, which made it unsafe for threading. The rewritten subroutine first precomputes the indices and stores them in an array, then does the loop over blocks which can now safely be threaded. The `single_PAO_to_grad` subroutine has not been refactored, it will be removed in a separate PR and the functionality merged with `single_PAO_to_grid` to address #244