@tkoskela 2023-06-13
I just realized the values in the output are only written out at 8 decimals. So my race condition is not producing any observable error. Very strange :thinking: