@davidbowler 2023-06-14
I have tested the attached file (216 atoms which scales happily enough to 9 MPI processes) on combinations from 1 to 8 MPI processes with 1 to 4 OpenMP threads and found the same energy to 10 decimal places for all cases.  I'm not sure if this helps!

[LiCltest.tgz](https://github.com/OrderN/CONQUEST-release/files/11745920/LiCltest.tgz)