@tkoskela 2023-08-24
I think this PR is now ready to merge to development. We can address #244 in a separate PR to avoid cluttering this one further. Here is a performance comparison to development. I ran these on young using 8 MPI ranks and the attached input files
[input_data.zip](https://github.com/OrderN/CONQUEST-release/files/12426913/input_data.zip). 

|   | develop | This PR  |
|---|---|---|
| Without `-fopenmp`                      | 95.331s  | 63.482s |
| With `-fopenmp`, 1 OMP thread   | 95.597s  | 62.324s |
| With `-fopenmp`, 4 OMP threads |   | 36.491 s |

Even the serial performance is significantly improved! I've started work on the code for #244 and it looks like we can gain a bit more speedup in threaded performance.