@davidbowler 2023-08-24
Running the LiCl test (input files above) on a local cluster (AMD, Intel compiler 2020_u4, MKL, OpenMPI 4.1.3) I get a segfault when using two threads and 9 MPI processes (the fault seems to come from a reduction).  Another segfault occurs using GCC 11.2.0 on the same system.  On a Mac MBP M2 with GCC 12.2 and OpenMPI 4.1.5 the same test runs without problems.