@davidbowler 2023-08-24
I find the same problem with your test above