@tkoskela 2023-08-24
Interesting, is it possible to get access to the cluster?

On young I'm using 
- `compilers/intel/2022.2`
- `mpi/intel/2021.6.0/intel`

gcc and openmpi are available there as well, I can see if I can build with them