@davidbowler 2023-08-24
The problem comes from `calc_matrix_elements_new` (found by trial and error - commenting out different areas) but I don't know more yet.