@davidbowler 2023-08-25
We are now confident that the problem here is coming from the OpenMP stack (the errors I found were caused by a stack overflow...).  Setting the environment variable OMP_STACKSIZE fixes this issue (in particular, I found that a value somewhere around 50M was helpful though this might require a little testing).

I will add to the documentation before approving the PR.